chore: restructure code - dashboard views - configs
